pyproject.toml
[tool.poetry]
name = "testfastapi"
version = "0.1.0"
description = ""
authors = ["ItSoda <nikitashchegilskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
python-decouple = "^3.8"
psycopg = "^3.1.18"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
uvicorn = "^0.29.0"


[tool.poetry.group.dev.dependencies]
flake8 = "5.0.4"
isort = "5.10.1"
black = "23.10.1"
coverage = "7.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"